==> 3-feature-extraction.ipynb <==
# **3. Feature Extraction**

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

data = pd.read_csv('./data/cleaned_chess_games.csv')
data.head(3)

import math

def organize_features(df):

    df_cpy = df.copy()
    
    # OPC
    # White openings
    WC1 = ['A00']
    WC2 = ['A40', 'B01', 'C00']
    WC3 = ['B00', 'C41', 'D00', 'C20']

    # Black openings
    BC1 = ['A00']
    BC2 = ['A40', 'B01', 'C00']
    BC3 = ['B00', 'C41', 'D00', 'C20']

    def w_opc(eco):
        if eco in WC1: return "WC1"
        elif eco in WC2: return "WC2"
        elif eco in WC3: return "WC3"
        else: return "N"
    
    def b_opc(eco):
        if eco in BC1: return "BC1"
        elif eco in BC2: return "BC2"
        elif eco in BC3: return "BC3"
        else: return "N"

    df_cpy['W_OPC'] = df_cpy['ECO'].apply(w_opc)
    df_cpy['B_OPC'] = df_cpy['ECO'].apply(b_opc)

    # WS_OPC, BS_OPC
    def ws_opc(whiteElo, blackElo, w_opc):
        if whiteElo >= blackElo:
            return f"1-{w_opc}"
        else:
            return f"0-{w_opc}"
    
    def bs_opc(whiteElo, blackElo, b_opc):
        if blackElo >= whiteElo:
            return f"1-{b_opc}"
        else:
            return f"0-{b_opc}"
    
    elo_opc = df_cpy[["WhiteElo", "BlackElo", "W_OPC", "B_OPC"]]

    elo_opc = elo_opc.assign(WS_OPC=elo_opc.apply(lambda row: ws_opc(row["WhiteElo"], row["BlackElo"], row["W_OPC"]), axis=1))
    elo_opc = elo_opc.assign(BS_OPC=elo_opc.apply(lambda row: bs_opc(row["WhiteElo"], row["BlackElo"], row["B_OPC"]), axis=1))

    df_cpy = pd.concat([df_cpy, elo_opc[["WS_OPC", "BS_OPC"]]], axis=1)

    df_cpy = df_cpy[['Event', 'W_ES', 'B_ES', 
                     'ECO', 'WS_OPC', 'BS_OPC',
                     'PG-MovesCount', 'White-Mistakes', 'Black-Mistakes', 'W-WP', 'B-WP',
                     'AN', 'W', 'B', 'Termination']]
    
    return df_cpy


games_data = organize_features(data)
games_data.head(3)

games_data.to_csv('./data/games_extracted_data.csv', index=False)
