==> 5-logistic-regression.ipynb <==
# **5. Logistic Regression**

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report,confusion_matrix

wwcX_train=pd.read_csv('train-test-split\white-win-loss\wwcX_train.csv')
wwcX_test=pd.read_csv('train-test-split\white-win-loss\wwcX_test.csv')
wwcY_train=pd.read_csv('train-test-split\white-win-loss\wwcY_train.csv')
wwcY_test=pd.read_csv('train-test-split\white-win-loss\wwcY_test.csv')

wtcX_train=pd.read_csv('train-test-split\white-termination\wtcX_train.csv')
wtcX_test=pd.read_csv('train-test-split\white-termination\wtcX_test.csv')
wtcY_train=pd.read_csv('train-test-split\white-termination\wtcY_train.csv')
wtcY_test=pd.read_csv('train-test-split\white-termination\wtcY_test.csv')

bwcX_train=pd.read_csv('train-test-split/black-win-loss/bwcX_train.csv')
bwcX_test=pd.read_csv('train-test-split/black-win-loss/bwcX_test.csv')
bwcY_train=pd.read_csv('train-test-split/black-win-loss/bwcY_train.csv')
bwcY_test=pd.read_csv('train-test-split/black-win-loss/bwcY_test.csv')

btcX_train=pd.read_csv('train-test-split/black-termination/btcX_train.csv')
btcX_test=pd.read_csv('train-test-split/black-termination/btcX_test.csv')
btcY_train=pd.read_csv('train-test-split/black-termination/btcY_train.csv')
btcY_test=pd.read_csv('train-test-split/black-termination/btcY_test.csv')

## **White**

### 1) White Win/Loss Classification

log_wwc=LogisticRegression()

log_wwc.fit(wwcX_train, wwcY_train)
wwcY_pred = log_wwc.predict(wwcX_test)

print("White Win/Loss Classification Report:")
print(classification_report(wwcY_test, wwcY_pred))
print('\n')
print(confusion_matrix(wwcY_test, wwcY_pred))

### 2) White Termination Classification

### White Termination Classification
log_wtc = LogisticRegression()
log_wtc.fit(wtcX_train, wtcY_train)
wtcY_pred = log_wtc.predict(wtcX_test)

print("White Termination Classification Report:")
print(classification_report(wtcY_test, wtcY_pred))
print('\n')
print(confusion_matrix(wtcY_test, wtcY_pred))

## **Black**

### Black Win/Loss Classification
log_bwc = LogisticRegression()
log_bwc.fit(bwcX_train, bwcY_train)
bwcY_pred = log_bwc.predict(bwcX_test)

print("Win/Loss Classification Report:")
print(classification_report(bwcY_test, bwcY_pred))
print('\n')
print(confusion_matrix(bwcY_test, bwcY_pred))

### Black Termination Classification
log_btc = LogisticRegression()
log_btc.fit(btcX_train, btcY_train)
btcY_pred = log_btc.predict(btcX_test)

print("Termination Classification Report:")
print(classification_report(btcY_test, btcY_pred))
print('\n')
print(confusion_matrix(btcY_test, btcY_test))
