==> 6-random-forests.ipynb <==
# **6. Random Forests**

import pandas as pd
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report, f1_score, precision_score, recall_score, accuracy_score, confusion_matrix

def rf_classification(X_train, y_train, X_test, y_test, mode):
    # Split the training data into training and validation sets
    X_train, X_val, y_train, y_val = train_test_split(
        X_train, y_train, test_size=0.2, random_state=42)

    # Define the parameter grid to search over
    param_grid = {
        'n_estimators': [100, 200],  # Number of trees in the forest
        'max_depth': [20, 30],  # Maximum depth of the tree
    }

    # Create a Random Forest classifier object
    rf_model = RandomForestClassifier(n_jobs=-1)

    # Create a GridSearchCV object
    if mode == "BINARY":
        grid_search = GridSearchCV(
            rf_model, param_grid, scoring='f1', cv=3, refit=True, verbose=3)
    elif mode == "MULTICLASS":
        grid_search = GridSearchCV(
            rf_model, param_grid, scoring='f1_micro', cv=3, refit=True, verbose=3)

    # Fit the GridSearchCV object to the data
    grid_search.fit(X_train, y_train)

    # Print the best hyperparameters and score
    print("Best hyperparameters: ", grid_search.best_params_)
    print("Best score: ", grid_search.best_score_)

    # Use the model to make predictions on the test data
    y_pred = grid_search.predict(X_test)

    # Get the precision and recall scores
    precision = precision_score(y_test, y_pred, average='weighted')
    recall = recall_score(y_test, y_pred, average='weighted')
    print("Precision:", precision)
    print("Recall:", recall)

    # Get the F1 score
    f1 = f1_score(y_test, y_pred, average='weighted')  # Use 'weighted' if you have imbalanced classes
    print("F1 Score:", f1)

    # Get the classification report
    report = classification_report(y_test, y_pred)
    print("Classification Report:\n", report)

def evaluation(y_test, y_pred, type, message):

    accuracy = accuracy_score(y_test, y_pred)
    conf_matrix = confusion_matrix(y_test, y_pred)

    if type == 'BINARY':
        precision = precision_score(y_test, y_pred)
        recall = recall_score(y_test, y_pred)
        f1 = f1_score(y_test, y_pred)
    
    elif type == 'MULTICLASS':
        precision = precision_score(y_test, y_pred, average='weighted')
        recall = recall_score(y_test, y_pred, average='weighted')
        f1 = f1_score(y_test, y_pred, average='weighted')
    
    print(f"{message}:")
    print("Accuracy: {:.4f}".format(accuracy))
    print("Precision: {:.4f}".format(precision))
    print("Recall: {:.4f}".format(recall))
    print("F1 Score: {:.4f}".format(f1))
    print("Confusion Matrix: \n{}\n".format(conf_matrix))


wwcX_train = pd.read_csv('./train-test-split/white-win-loss/wwcX_train.csv')
wwcX_test = pd.read_csv('./train-test-split/white-win-loss/wwcX_test.csv')
wwcY_train = pd.read_csv('./train-test-split/white-win-loss/wwcY_train.csv')
wwcY_test = pd.read_csv('./train-test-split/white-win-loss/wwcY_test.csv')

wtcX_train = pd.read_csv('./train-test-split/white-termination/wtcX_train.csv')
wtcX_test = pd.read_csv('./train-test-split/white-termination/wtcX_test.csv')
wtcY_train = pd.read_csv('./train-test-split/white-termination/wtcY_train.csv')
wtcY_test = pd.read_csv('./train-test-split/white-termination/wtcY_test.csv')

bwcX_train = pd.read_csv('./train-test-split/black-win-loss/bwcX_train.csv')
bwcX_test = pd.read_csv('./train-test-split/black-win-loss/bwcX_test.csv')
bwcY_train = pd.read_csv('./train-test-split/black-win-loss/bwcY_train.csv')
bwcY_test = pd.read_csv('./train-test-split/black-win-loss/bwcY_test.csv')

btcX_train = pd.read_csv('./train-test-split/black-termination/btcX_train.csv')
btcX_test = pd.read_csv('./train-test-split/black-termination/btcX_test.csv')
btcY_train = pd.read_csv('./train-test-split/black-termination/btcY_train.csv')
btcY_test = pd.read_csv('./train-test-split/black-termination/btcY_test.csv')

## **White**

### White Win/Loss Classification
rf_wwc = RandomForestClassifier(n_estimators=100, random_state=42)
rf_wwc.fit(wwcX_train, wwcY_train)
wwcY_pred = rf_wwc.predict(wwcX_test)

print("White Win/Loss Classification Report:")
print(classification_report(wwcY_test, wwcY_pred))

### White Termination Classification
rf_wtc = RandomForestClassifier(n_estimators=100, random_state=42)
rf_wtc.fit(wtcX_train, wtcY_train)
wtcY_pred = rf_wtc.predict(wtcX_test)

print("White Termination Classification Report:")
print(classification_report(wtcY_test, wtcY_pred))

### White Win/Loss Classification
evaluation(wwcY_test, wwcY_pred, "BINARY", "White Win/Loss Classification Report")

### White Termination Classification
evaluation(wtcY_test, wtcY_pred, "MULTICLASS", "White Termination Classification Report")

## **Black**

### Black Win/Loss Classification
rf_bwc = RandomForestClassifier(n_estimators=100, random_state=42)
rf_bwc.fit(bwcX_train, bwcY_train)
bwcY_pred = rf_bwc.predict(bwcX_test)

print("Win/Loss Classification Report:")
print(classification_report(bwcY_test, bwcY_pred))

### Black Termination Classification
rf_btc = RandomForestClassifier(n_estimators=100, random_state=42)
rf_btc.fit(btcX_train, btcY_train)
btcY_pred = rf_btc.predict(btcX_test)

print("Termination Classification Report:")
print(classification_report(btcY_test, btcY_pred))

### Black Win/Loss Classification
evaluation(bwcY_test, bwcY_pred, "BINARY", "Black Win/Loss Classification Report")

### Black Termination Classification
evaluation(btcY_test, btcY_pred, "MULTICLASS", "Black Termination Classification Report")

# Hyperparameter Tuning for Random Forests

## **White**

### 1) White Win/Loss Classification

print("White Win-Loss Classification")
rf_classification(wwcX_train, wwcY_train['W'], wwcX_test, wwcY_test['W'], "BINARY")

### 2) White Termination Classification

print("White Termination Classification")
rf_classification(wtcX_train, wtcY_train['Termination'], wtcX_test, wtcY_test['Termination'], "MULTICLASS")

## **Black**

### 1) Black Win/Loss Classification

print("Black Win-Loss Classification")
rf_classification(bwcX_train, bwcY_train['B'], bwcX_test, bwcY_test['B'], "BINARY")

### 2) Black Termination Classification

print("Black Termination Classification")
rf_classification(btcX_train, btcY_train['Termination'], btcX_test, btcY_test['Termination'], "MULTICLASS")

Actually Win/Loss Classification performance got lower after hyperparameter tuning, but Termination Classification performance has improved.
